[Food dialog update] + Added saveChanges logic, the item will be updated but not with persistence + Added `getValue` methods to Row objects in order to be able to fetch their new value + Changed food values from int to double

+++ lib/persistence/entities/pantry_item.dart
@@ -24,7 +24,7 @@ class PantryItem {
   @ignore
   Food food;
 
-  int leftQuantity;
+  double leftQuantity;
 
   PantryItem({this.id, this.foodId, this.food, this.leftQuantity});
 }
